# swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API de gerenciamento de vendas utilizando arquitetura MSC",
    "description": "Ojetivo foi criar um sistema de gerenciamento de vendas utilizando arquitetura MSC",
    "contact": {
      "email": "pabloalmeidacti@gmail.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags":[
    {
      "name": "product",
      "description": "Tudo sobre produtos"
    },
    {
      "name": "sale",
      "description": "Tudo sobre vendas"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/products/": {
      "get": {
        "tags": [
          "product"
        ],
        "summary":"Lista todos os produtos",
        "description": "Rota responsável por listar todos os produtos",
        "parameters": [],
        "responses": {
          "200": {
            "$ref": "#/definitions/responses/OK",
            "schema": {
              "$ref": "#/definitions/schemas/AllProducts"
            }
          },
          "404": {
            "$ref": "#/definitions/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "product"
          ],
          "summary":"Adiciona um novo produto",
          "description": "Rota responsável por adicionar um novo produto",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/schemas/NewProduct"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/definitions/responses/OK",
            "schema": {
              "$ref": "#/definitions/schemas/Product"
            }
          },
          "409": {
            "$ref": "#/definitions/responses/InvalidBody"
          },
          "400": {
            "$ref": "#/definitions/responses/InvalidBody"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "product"
          ],
          "summary":"Lista um produto pelo ID",
          "description": "Rota responsável por listar um produto pelo seu ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/responses/OK",
            "schema": {
              "$ref": "#/definitions/schemas/Product"
            }
          },
          "404": {
            "$ref": "#/definitions/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "product"
          ],
          "summary":"Atualiza um produto pelo ID",
          "description": "Rota responsável por atualizar um produto pelo seu ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/schemas/NewProduct"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/definitions/responses/OK",
            "schema": {
              "$ref": "#/definitions/schemas/Product"
            }
          },
          "409": {
            "$ref": "#/definitions/responses/InvalidBody"
          },
          "400": {
            "$ref": "#/definitions/responses/InvalidBody"
          }
        }
      },
      "delete": {
        "tags": [
          "product"
          ],
          "summary":"Deleta um produto pelo ID",
          "description": "Rota responsável por deletar um produto pelo seu ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Sem conteúdo"
          },
          "404": {
            "$ref": "#/definitions/responses/NotFound"
          }
        }
      }
    },
    "/sales/": {
      "get": {
        "tags": [
          "sale"
          ],
          "summary":"Lista todas as vendas",
          "description": "Rota responsável por listar todas as vendas",
        "parameters": [],
        "responses": {
          "200": {
            "$ref": "#/definitions/responses/OK",
            "schema": {
              "$ref": "#/definitions/schemas/AllSales"
            }
          },
          "404": {
            "$ref": "#/definitions/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "sale"
          ],
          "summary":"Adiciona uma nova venda",
          "description": "Rota responsável por adicionar uma nova venda",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/schemas/NewSale"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/definitions/responses/OK",
            "schema": {
              "$ref": "#/definitions/schemas/Sale"
            }
          },
          "409": {
            "$ref": "#/definitions/responses/InvalidBody"
          },
          "400": {
            "$ref": "#/definitions/responses/InvalidBody"
          }
        }
      }
    },
    "/sales/{id}": {
      "get": {
        "tags": [
          "sale"
          ],
          "summary":"Lista uma venda pelo ID",
          "description": "Rota responsável por listar uma venda pelo seu ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/responses/OK",
            "schema": {
              "$ref": "#/definitions/schemas/AllSales"
            }
          },
          "404": {
            "$ref": "#/definitions/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "sale"
          ],
          "summary":"Deleta uma venda pelo ID",
          "description": "Rota responsável por deletar uma venda pelo seu ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Sem conteúdo"
          },
          "404": {
            "$ref": "#/definitions/responses/NotFound"
          }
        }
      }
    }
  },
  "definitions": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name":{
            "type": "string",
            "example": "produto A"
          },
          "quantity": {
            "type": "number",
            "example": 10
          }
        }
      },
      "NewProduct": {
        "type": "object",
        "properties": {
          "name":{
            "type": "string",
            "example": "produto A"
          },
          "quantity": {
            "type": "number",
            "example": 10
          }
        }
      }, 
      "AllProducts": {
        "type":"array",
        "items": {
          "$ref": "#/definitions/schemas/Product"
        }
      },
      "Sales": {
        "type":"object",
        "properties": {
          "saleId": {
            "type":"number",
            "example": 1
          },
          "date": {
            "type":"string",
            "example": "2021-09-09T04:54:29.000Z"
          },
          "productId": {
            "type":"number",
            "example": 1
          },
          "quantity": {
            "type":"number",
            "example": 2
          }
        }
      },
      "AllSales": {
        "type":"array",
        "items": {
          "$ref": "#/definitions/schemas/Sales"
        }
      },
      "NewSale": {
        "type":"array",
        "properties": {
          "productId": {
            "type":"number",
            "example": 1
          },
          "quantity": {
            "type":"number",
            "example": 3
          }
        }
      },
      "UpdateSale": {
        "type":"array",
        "items": {
          "$ref": "#/definitions/schemas/NewSale"
        }
      },
      "NewSaleResponse": {
        "type":"object",
        "properties": {
          "id": {
            "type":"number",
            "example": 1
          },
          "itemsSold": {
            "type":"array",
            "example": {
              "productId": 1,
              "quantity": 3
            }
          }
        }
      },
      "UpdateSaleResponse": {
        "type":"object",
        "properties": {
          "saleId": {
            "type":"number",
            "example": 1
          },
          "itemsUpdated": {
            "type":"array",
            "example": {
              "productId": 1,
              "quantity": 3
            }
          }
        }
      }
    },
    "responses": {
      "OK": {
        "description": "Requisição bem sucedida"
      },
      "NotFound": {
        "description": "Informação não encontrada"
      },
      "InvalidBody": {
        "description": "Corpo da requisição inválido"
      }
    }
  }
}